// client/src/providers/apiClient/model/friendInput.ts
/**
 * vinimay
 * Vinimay is a decentralised social network focused on giving back control of its data to the user
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface FriendInput {
    /**
     * Request recipient
     */
    to: string;

    /**
     * Type of request
     */
    type: FriendInput.TypeEnum;

}
export namespace FriendInput {
    export enum TypeEnum {
        Friend = <any> 'friend',
        Following = <any> 'following'
    }
}